+++ src/VsCodeDialog/vscode.ts
@@ -6,6 +6,7 @@
 
 import { spawn } from 'child_process';
 import { remote } from 'electron';
+import os from 'os';
 import { logger, usageData } from 'pc-nrfconnect-shared';
 
 import { Dispatch, RootState } from '../state';
@@ -47,6 +48,15 @@ const EXTENSIONS = [
     { id: 'twxs.cmake', name: 'CMake' },
 ];
 
+export enum NrfjprogStatus {
+    NOT_INSTALLED,
+    INSTALLED,
+    M1_VERSION,
+}
+
+export const isAppleSilicon =
+    process.platform === 'darwin' && os.cpus()[0].model.includes('Apple');
+
 const minDelay = 500;
 export const checkOpenVsCodeWithDelay = () => (dispatch: Dispatch) => {
     dispatch(setVsCodeStatus(VsCodeStatus.NOT_CHECKED));
@@ -80,7 +90,16 @@ export const getVsCodeStatus = () => async (dispatch: Dispatch) => {
             status = VsCodeStatus.MISSING_EXTENSIONS;
         else {
             const nrfjprog = await getNrfjprogStatus();
-            if (!nrfjprog) status = VsCodeStatus.MISSING_NRFJPROG;
+            if (isAppleSilicon) {
+                const vscode = await spawnAsync(
+                    'file "$(dirname "$(readlink $(which code))")/../../../MacOS/Electron"'
+                );
+                if ((await checkExecArchitecture(vscode)) !== 'x86_64')
+                    status = VsCodeStatus.INSTALL_INTEL;
+            } else if (nrfjprog === NrfjprogStatus.NOT_INSTALLED)
+                status = VsCodeStatus.MISSING_NRFJPROG;
+            else if (nrfjprog === NrfjprogStatus.M1_VERSION)
+                status = VsCodeStatus.NRFJPROG_INSTALL_INTEL;
             else status = VsCodeStatus.INSTALLED;
         }
     } catch {
@@ -99,7 +118,7 @@ export const installExtensions =
 const installExtension = (identifier: string) => async (dispatch: Dispatch) => {
     try {
         dispatch(startInstallingExtension(identifier));
-        await spawnAsync(['--install-extension', identifier]);
+        await spawnAsync('code', ['--install-extension', identifier]);
         dispatch(installedExtension(identifier));
         logger.info(`Installed extension ${identifier}`);
     } catch {
@@ -109,7 +128,11 @@ const installExtension = (identifier: string) => async (dispatch: Dispatch) => {
 };
 
 export const listInstalledExtensions = async (): Promise<VsCodeExtension[]> => {
-    const installedExtensions = await spawnAsync(['--list-extensions']);
+    const installedExtensions = (
+        await spawnAsync('code', ['--list-extensions'])
+    )
+        .trim()
+        .split('\n');
     return EXTENSIONS.map(extension => ({
         identifier: extension.id,
         name: extension.name,
@@ -119,23 +142,22 @@ export const listInstalledExtensions = async (): Promise<VsCodeExtension[]> => {
     }));
 };
 
-export const getNrfjprogStatus = () => {
-    return new Promise<boolean>(resolve => {
-        const spawnAsync = spawn('nrfjprog', {
-            shell: true,
-            env: {
-                ...remote.process.env,
-                PATH: pathEnvVariable(),
-            },
-        });
-
-        spawnAsync.on('close', (code, signal) => {
-            if (code === 0 && signal === null) {
-                return resolve(true);
+export const getNrfjprogStatus = async () => {
+    try {
+        await spawnAsync('nrfjprog');
+        try {
+            if (isAppleSilicon) {
+                const stdout = await spawnAsync('file $(which jlinkexe)');
+                if (checkExecArchitecture(stdout) !== 'x86_64')
+                    return NrfjprogStatus.M1_VERSION;
             }
-            return resolve(false);
-        });
-    });
+        } catch {
+            return NrfjprogStatus.M1_VERSION;
+        }
+        return NrfjprogStatus.INSTALLED;
+    } catch {
+        return NrfjprogStatus.NOT_INSTALLED;
+    }
 };
 
 const pathEnvVariable = () => {
@@ -144,9 +166,17 @@ const pathEnvVariable = () => {
     return `/usr/local/bin:${remote.process.env.PATH}`;
 };
 
-const spawnAsync = (params: string[]) => {
-    return new Promise<string[]>((resolve, reject) => {
-        const codeProcess = spawn('code', params, {
+const checkExecArchitecture = (stdout: string) => {
+    const universalMatch = 'Mach-O universal binary with 2 architectures';
+    const intelMatch = 'Mach-O 64-bit executable x86_64';
+    if (stdout.includes(universalMatch)) return 'universal';
+    if (stdout.includes(intelMatch)) return 'x86_64';
+    return 'arm';
+};
+
+const spawnAsync = (cmd: string, params?: string[]) => {
+    return new Promise<string>((resolve, reject) => {
+        const codeProcess = spawn(cmd, params, {
             shell: true,
             env: {
                 ...remote.process.env,
@@ -164,7 +194,7 @@ const spawnAsync = (params: string[]) => {
 
         codeProcess.on('close', (code, signal) => {
             if (code === 0 && signal === null) {
-                return resolve(stdout.trim().split('\n'));
+                return resolve(stdout);
             }
             if (stderr) console.log(stderr);
             return reject();
@@ -174,5 +204,5 @@ const spawnAsync = (params: string[]) => {
 
 export const openVsCode = () => {
     usageData.sendUsageData(EventAction.OPEN_VS_CODE, process.platform);
-    spawnAsync([]);
+    spawnAsync('code');
 };